+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The major version bump is due to dropped support for PHP `7.2` and a several bre
 * Added `Collection::first()` method.
 * Added `Collection::last()` method.
 * Added support for the `samesite` cookie option (defaults to `Lax`).
+* Added `Query::withCountOf()` method to the ORM query builder.
 
 #### Changes